% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmp.R
\name{dcmp}
\alias{dcmp}
\title{The Conway-Maxwell Poisson Probability Mass Function}
\usage{
dcmp(x, mu, phi, max_terms = 1000, tol = 1e-12)
}
\arguments{
\item{x}{A vector of non-negative integers.}

\item{mu}{A numeric value representing the location parameter \eqn{\mu}. Must be positive.}

\item{phi}{A numeric value representing the dispersion parameter \eqn{\phi}.}

\item{max_terms}{An integer for the maximum number of terms to sum for the normalizing constant.}

\item{tol}{A numeric tolerance for the normalizing constant sum. Summation stops when
the ratio of the current term to the total sum falls below this tolerance.}
}
\value{
A numeric vector of probabilities.
}
\description{
Computes the probability mass function (PMF) of the Conway-Maxwell Poisson (mCMP) distribution.
The PMF is given by:
\deqn{ p(X=x;\mu,\phi) = \frac{\left( \mu + \frac{e^\phi - 1}{2e^\phi} \right)^{x e^\phi}}{(x!)^{e^\phi} Z(\mu,\phi)} }
where \eqn{x = 0, 1, 2, \ldots}, \eqn{\mu > 0}, \eqn{-\infty < \phi < \infty}, and \eqn{Z(\mu,\phi)} is the normalizing constant.
The normalizing constant is defined as:
\deqn{ Z(\mu,\phi) = \sum_{j=0}^{\infty} \frac{\left( \mu + \frac{e^\phi - 1}{2e^\phi} \right)^{j e^\phi}}{(j!)^{e^\phi}} }
This function uses an iterative summation for \eqn{Z(\mu,\phi)} up to a specified number of \code{max_terms}
or until the added terms fall below a given \code{tol}.
}
\examples{
dcmp(0:5, mu = 2, phi = 0) # Equi-dispersed (Poisson)
dcmp(0:5, mu = 2, phi = 1) # Under-dispersed
dcmp(0:5, mu = 2, phi = -1) # Over-dispersed
}
